Front Page README
A README that contains an overview of the project and a members list. I am not sure of Ryan R's surname. Everyone can add their own github links after your talk tomorrow. It will force them to get an account, learn how to fork, edit (and md), and then push.